Fix rotation tests not being executed
