+++ pyresample/test/test_utils.py
@@ -1,17 +1,29 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+# Copyright (c) 2015-2019 Pyresample developers
+#
+# This program is free software: you can redistribute it and/or modify it under
+# the terms of the GNU Lesser General Public License as published by the Free
+# Software Foundation, either version 3 of the License, or (at your option) any
+# later version.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+# FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License for more
+# details.
+#
+# You should have received a copy of the GNU Lesser General Public License along
+# with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""Test various utility functions."""
+
 import os
 import unittest
 
 import numpy as np
 import uuid
 
-import pyresample.utils._proj4
-import pyresample.utils._rasterio
 from pyresample.test.utils import create_test_longitude, create_test_latitude
-
-
-def tmp(f):
-    f.tmp = True
-    return f
 
 
 def tmptiff(width=100, height=100, transform=None, crs=None, dtype=np.uint8):
@@ -167,7 +179,7 @@ class TestPreprocessing(unittest.TestCase):
     def test_nearest_neighbor_area_area(self):
         from pyresample import utils, geometry
         proj_str = "+proj=lcc +datum=WGS84 +ellps=WGS84 +lat_0=25 +lat_1=25 +lon_0=-95 +units=m +no_defs"
-        proj_dict = pyresample.utils._proj4.proj4_str_to_dict(proj_str)
+        proj_dict = utils._proj4.proj4_str_to_dict(proj_str)
         extents = [0, 0, 1000. * 5000, 1000. * 5000]
         area_def = geometry.AreaDefinition('CONUS', 'CONUS', 'CONUS',
                                            proj_dict, 400, 500, extents)
@@ -184,7 +196,7 @@ class TestPreprocessing(unittest.TestCase):
         grid = geometry.GridDefinition(lons=lon_arr, lats=lat_arr)
 
         proj_str = "+proj=lcc +datum=WGS84 +ellps=WGS84 +lat_0=25 +lat_1=25 +lon_0=-95 +units=m +no_defs"
-        proj_dict = pyresample.utils._proj4.proj4_str_to_dict(proj_str)
+        proj_dict = utils._proj4.proj4_str_to_dict(proj_str)
         extents = [0, 0, 1000. * 5000, 1000. * 5000]
         area_def = geometry.AreaDefinition('CONUS', 'CONUS', 'CONUS',
                                            proj_dict, 400, 500, extents)
@@ -193,7 +205,7 @@ class TestPreprocessing(unittest.TestCase):
     def test_nearest_neighbor_grid_area(self):
         from pyresample import utils, geometry
         proj_str = "+proj=lcc +datum=WGS84 +ellps=WGS84 +lat_0=25 +lat_1=25 +lon_0=-95 +units=m +no_defs"
-        proj_dict = pyresample.utils._proj4.proj4_str_to_dict(proj_str)
+        proj_dict = utils._proj4.proj4_str_to_dict(proj_str)
         extents = [0, 0, 1000. * 2500., 1000. * 2000.]
         area_def = geometry.AreaDefinition('CONUS', 'CONUS', 'CONUS',
                                            proj_dict, 40, 50, extents)
@@ -255,7 +267,7 @@ class TestMisc(unittest.TestCase):
 
     def test_proj4_radius_parameters_provided(self):
         from pyresample import utils
-        a, b = pyresample.utils._proj4.proj4_radius_parameters(
+        a, b = utils._proj4.proj4_radius_parameters(
             '+proj=stere +a=6378273 +b=6356889.44891',
         )
         np.testing.assert_almost_equal(a, 6378273)
@@ -263,7 +275,7 @@ class TestMisc(unittest.TestCase):
 
     def test_proj4_radius_parameters_ellps(self):
         from pyresample import utils
-        a, b = pyresample.utils._proj4.proj4_radius_parameters(
+        a, b = utils._proj4.proj4_radius_parameters(
             '+proj=stere +ellps=WGS84',
         )
         np.testing.assert_almost_equal(a, 6378137.)
@@ -271,7 +283,7 @@ class TestMisc(unittest.TestCase):
 
     def test_proj4_radius_parameters_default(self):
         from pyresample import utils
-        a, b = pyresample.utils._proj4.proj4_radius_parameters(
+        a, b = utils._proj4.proj4_radius_parameters(
             '+proj=lcc',
         )
         # WGS84
@@ -281,9 +293,9 @@ class TestMisc(unittest.TestCase):
     def test_proj4_str_dict_conversion(self):
         from pyresample import utils
         proj_str = "+proj=lcc +ellps=WGS84 +lon_0=-95 +no_defs"
-        proj_dict = pyresample.utils._proj4.proj4_str_to_dict(proj_str)
-        proj_str2 = pyresample.utils._proj4.proj4_dict_to_str(proj_dict)
-        proj_dict2 = pyresample.utils._proj4.proj4_str_to_dict(proj_str2)
+        proj_dict = utils._proj4.proj4_str_to_dict(proj_str)
+        proj_str2 = utils._proj4.proj4_dict_to_str(proj_dict)
+        proj_dict2 = utils._proj4.proj4_str_to_dict(proj_str2)
         self.assertDictEqual(proj_dict, proj_dict2)
         self.assertIsInstance(proj_dict['lon_0'], float)
         self.assertIsInstance(proj_dict2['lon_0'], float)
@@ -306,6 +318,7 @@ class TestMisc(unittest.TestCase):
             os.remove(yaml_file)
 
     def test_get_area_def_from_raster(self):
+        from pyresample import utils
         from rasterio.crs import CRS
         from affine import Affine
         x_size = 791
@@ -317,7 +330,7 @@ class TestMisc(unittest.TestCase):
         area_id = 'area_id'
         proj_id = 'proj_id'
         name = 'name'
-        area_def = pyresample.utils._rasterio.get_area_def_from_raster(
+        area_def = utils._rasterio.get_area_def_from_raster(
             source, area_id=area_id, name=name, proj_id=proj_id)
         self.assertEqual(area_def.area_id, area_id)
         self.assertEqual(area_def.proj_id, proj_id)
@@ -333,7 +346,7 @@ class TestMisc(unittest.TestCase):
         from pyresample import utils
         crs = CRS(init='epsg:3857')
         source = tmptiff(crs=crs)
-        area_def = pyresample.utils._rasterio.get_area_def_from_raster(source)
+        area_def = utils._rasterio.get_area_def_from_raster(source)
         self.assertEqual(area_def.proj_id, 'WGS 84 / Pseudo-Mercator')
 
     def test_get_area_def_from_raster_rotated_value_err(self):
@@ -342,7 +355,7 @@ class TestMisc(unittest.TestCase):
         transform = Affine(300.0379266750948, 0.1, 101985.0,
                            0.0, -300.041782729805, 2826915.0)
         source = tmptiff(transform=transform)
-        self.assertRaises(ValueError, pyresample.utils._rasterio.get_area_def_from_raster, source)
+        self.assertRaises(ValueError, utils._rasterio.get_area_def_from_raster, source)
 
     def test_get_area_def_from_raster_non_georef_value_err(self):
         from pyresample import utils
@@ -350,7 +363,7 @@ class TestMisc(unittest.TestCase):
         transform = Affine(300.0379266750948, 0.0, 101985.0,
                            0.0, -300.041782729805, 2826915.0)
         source = tmptiff(transform=transform)
-        self.assertRaises(ValueError, pyresample.utils._rasterio.get_area_def_from_raster, source)
+        self.assertRaises(ValueError, utils._rasterio.get_area_def_from_raster, source)
 
     def test_get_area_def_from_raster_non_georef_respects_proj_dict(self):
         from pyresample import utils
@@ -359,8 +372,108 @@ class TestMisc(unittest.TestCase):
                            0.0, -300.041782729805, 2826915.0)
         source = tmptiff(transform=transform)
         proj_dict = {'init': 'epsg:3857'}
-        area_def = pyresample.utils._rasterio.get_area_def_from_raster(source, proj_dict=proj_dict)
+        area_def = utils._rasterio.get_area_def_from_raster(source, proj_dict=proj_dict)
         self.assertDictEqual(area_def.proj_dict, proj_dict)
+
+
+class TestProjRotation(unittest.TestCase):
+    """Test loading areas with rotation specified."""
+
+    def test_rotation_legacy(self):
+        """Basic rotation in legacy format."""
+        from pyresample.area_config import load_area
+        legacyDef = """REGION: regionB {
+        NAME:          regionB
+        PCS_ID:        regionB
+        PCS_DEF:       proj=merc, lon_0=-34, k=1, x_0=0, y_0=0, a=6378137, b=6378137
+        XSIZE:         800
+        YSIZE:         548
+        ROTATION:      -45
+        AREA_EXTENT:   (-7761424.714818418, -4861746.639279127, 11136477.43264252, 8236799.845095873)
+        };"""
+        flegacy = "/tmp/TestProjRotation_test_rotation_legacy.txt"
+        f = open(flegacy, "w")
+        f.write(legacyDef)
+        f.close()
+        test_area = load_area(flegacy, 'regionB')
+        self.assertEqual(test_area.rotation, -45)
+        os.remove(flegacy)
+
+    def test_rotation_yaml(self):
+        """Basic rotation in yaml format."""
+        from pyresample.area_config import load_area
+        yamlDef = """regionB:
+          description: regionB
+          projection:
+            a: 6378137.0
+            b: 6378137.0
+            lon_0: -34
+            proj: merc
+            x_0: 0
+            y_0: 0
+            k_0: 1
+          shape:
+            height: 548
+            width: 800
+          rotation: -45
+          area_extent:
+            lower_left_xy: [-7761424.714818418, -4861746.639279127]
+            upper_right_xy: [11136477.43264252, 8236799.845095873]
+          units: m"""
+        fyaml = "/tmp/TestProjRotation_test_rotation_yaml.txt"
+        f = open(fyaml,"w")
+        f.write(yamlDef)
+        f.close()
+        test_area = load_area(fyaml, 'regionB')
+        self.assertEqual(test_area.rotation, -45)
+        os.remove(fyaml)
+
+    def test_norotation_legacy(self):
+        """No rotation specified in legacy format."""
+        from pyresample.area_config import load_area
+        legacyDef = """REGION: regionB {
+        NAME:          regionB
+        PCS_ID:        regionB
+        PCS_DEF:       proj=merc, lon_0=-34, k=1, x_0=0, y_0=0, a=6378137, b=6378137
+        XSIZE:         800
+        YSIZE:         548
+        AREA_EXTENT:   (-7761424.714818418, -4861746.639279127, 11136477.43264252, 8236799.845095873)
+        };"""
+        flegacy = "/tmp/TestProjRotation_test_rotation_legacy.txt"
+        f = open(flegacy, "w")
+        f.write(legacyDef)
+        f.close()
+        test_area = load_area(flegacy, 'regionB')
+        self.assertEqual(test_area.rotation, 0)
+        os.remove(flegacy)
+
+    def test_norotation_yaml(self):
+        """No rotation specified in yaml format."""
+        from pyresample.area_config import load_area
+        yamlDef = """regionB:
+          description: regionB
+          projection:
+            a: 6378137.0
+            b: 6378137.0
+            lon_0: -34
+            proj: merc
+            x_0: 0
+            y_0: 0
+            k_0: 1
+          shape:
+            height: 548
+            width: 800
+          area_extent:
+            lower_left_xy: [-7761424.714818418, -4861746.639279127]
+            upper_right_xy: [11136477.43264252, 8236799.845095873]
+          units: m"""
+        fyaml = "/tmp/TestProjRotation_test_rotation_yaml.txt"
+        f = open(fyaml, "w")
+        f.write(yamlDef)
+        f.close()
+        test_area = load_area(fyaml, 'regionB')
+        self.assertEqual(test_area.rotation, 0)
+        os.remove(fyaml)
 
 
 def suite():
@@ -372,6 +485,7 @@ def suite():
     mysuite.addTest(loader.loadTestsFromTestCase(TestYAMLAreaParser))
     mysuite.addTest(loader.loadTestsFromTestCase(TestPreprocessing))
     mysuite.addTest(loader.loadTestsFromTestCase(TestMisc))
+    mysuite.addTest(loader.loadTestsFromTestCase(TestProjRotation))
 
     return mysuite
